Renaming components and fixing a few typos.
