+++ apps/core/app/(landing)/_components/desktop-navbar/feather-navbar-authenticated.tsx
